feat: replaced terraform with cdk
moved lambda out of vpc
tested multiple times and deploy, destroy, and all gui interactions work
