+++ .gitignore
@@ -1,13 +1,10 @@
 venv
 .env
 python/*
-layer_content.zip
-.terraform
 create_layer
 **/*.pem
-terraform_overview.sh
-terraform_project_overview.txt
-terraform_output.txt
-*.tfstate
-*.tfstate.*
-terraform.tfstate
+node_modules/
+cdk.context.json
+cdk.out/
+.cdk.staging
+
