+++ .npmignore
@@ -1,0 +1,3 @@
+# CDK asset staging directory
+.cdk.staging
+cdk.out
