--- cdk_deploy/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
