--- cdk_deploy/.npmignore
@@ -1,3 +1,0 @@
-# CDK asset staging directory
-.cdk.staging
-cdk.out
