--- modules/clickhouse_ec2_instance/variables.tf
@@ -1,9 +1,0 @@
-variable "vpc_id" {
-  description = "The ID of the VPC"
-  type        = string
-}
-
-variable "subnet_id" {
-  description = "The ID of the subnet where the instance will be launched"
-  type        = string
-}
