+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "cdk_deploy",
+  "version": "0.1.0",
+  "bin": {
+    "cdk_deploy": "bin/cdk_deploy.js"
+  },
+  "scripts": {
+    "build": "echo \"The build step is not required when using JavaScript!\" && exit 0",
+    "cdk": "cdk",
+    "test": "jest"
+  },
+  "devDependencies": {
+    "aws-cdk": "2.149.0",
+    "jest": "^29.7.0"
+  },
+  "dependencies": {
+    "@aws-cdk/aws-secretsmanager": "^1.204.0",
+    "aws-cdk-lib": "^2.149.0",
+    "cdk-ec2-key-pair": "^4.0.1",
+    "constructs": "^10.0.0"
+  }
+}
