+++ test/cdk_deploy.test.js
@@ -1,0 +1,17 @@
+// const cdk = require('aws-cdk-lib');
+// const { Template } = require('aws-cdk-lib/assertions');
+// const CdkDeploy = require('../lib/cdk_deploy-stack');
+
+// example test. To run these tests, uncomment this file along with the
+// example resource in lib/cdk_deploy-stack.js
+test('SQS Queue Created', () => {
+//   const app = new cdk.App();
+//   // WHEN
+//   const stack = new CdkDeploy.CdkDeployStack(app, 'MyTestStack');
+//   // THEN
+//   const template = Template.fromStack(stack);
+
+//   template.hasResourceProperties('AWS::SQS::Queue', {
+//     VisibilityTimeout: 300
+//   });
+});
